+++ app/shell/py/pie/pie/update/__init__.py
@@ -1,0 +1,1 @@
+"""Metadata update helpers."""
