+++ app/shell/py/pie/setup.py
@@ -19,8 +19,8 @@ setup(
         'console_scripts': [
             'build-index=pie.build_index:main',
             'update-index=pie.update_index:main',
-            'update-pubdate=pie.update_pubdate:main',
-            'update-author=pie.update_author:main',
+            'update-pubdate=pie.update.pubdate:main',
+            'update-author=pie.update.author:main',
             'picasso=pie.picasso:main',
             'render-jinja-template=pie.render_jinja_template:main',
             'render-study-json=pie.render_study_json:main',
@@ -29,8 +29,8 @@ setup(
             'detect-html-dicts=pie.detect_html_dicts:main',
             'indextree-json=pie.indextree_json:main',
             'gen-markdown-index=pie.gen_markdown_index:main',
-            'check-page-title=pie.check_page_title:main',
-            'check-post-build=pie.check_post_build:main',
+            'check-page-title=pie.check.page_title:main',
+            'check-post-build=pie.check.post_build:main',
             'create-post=pie.create_post:main',
             'create=pie.create:main',
             'emojify=pie.emojify:main',
