+++ docs/reference/update-common.md
@@ -3,7 +3,8 @@
 Shared helpers for the metadata update scripts.
 
 The module provides utilities used by command line tools such as
-`update-author` and `update-pubdate`:
+`update-author` and `update-pubdate` (implemented in
+:mod:`pie.update.author` and :mod:`pie.update.pubdate`):
 
 - `get_changed_files()` – return a list of tracked files modified in git.
 - `replace_field(fp, field, value)` – replace a metadata field in YAML or
